Move open URL to header level
