+++ podcasts/Analytics/AnalyticsEvent.swift
@@ -306,6 +306,7 @@ enum AnalyticsEvent: String {
     case podcastsScreenSortOrderChanged
     case podcastsScreenEpisodeGroupingChanged
     case podcastsScreenTabTapped
+    case podcastScreenPodcastDescriptionLinkTapped
 
     // MARK: - App Store Review Request
